Refactor PBC wrapping logic in run_sim.py and pbc_utils_jax.py
- Changed data type for cell dimensions in `run_sim.py` from float64 to float32 to optimize memory usage.
- Enhanced the `wrap_groups` function in `pbc_utils_jax.py` for improved numerical stability by using a more precise method for shifting monomer groups, ensuring accurate handling of periodic boundary conditions.

+++ mmml/cli/run_sim.py
@@ -1145,7 +1145,7 @@ shake bonh para sele all end
             # PBC WRAPPING SETUP
             # ========================================================================
             if use_pbc:
-                _cell_jax = jnp.asarray(atoms.get_cell()[:], dtype=jnp.float64)
+                _cell_jax = jnp.asarray(atoms.get_cell()[:], dtype=jnp.float32)
                 _monomer_groups = [
                     jnp.arange(int(monomer_offsets[m]), int(monomer_offsets[m + 1]))
                     for m in range(n_monomers)
